Add season statistic module and moved season methods to said module. Delete league_statistics file.
